Use task queue for exporter
